assignment/serializers.py
from rest_framework import serializers
from .models import Assignment, AssignmentFile

class AssignmentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Assignment
        fields = '__all__'

class AssignmentFileSerializer(serializers.ModelSerializer):
    class Meta:
        model = AssignmentFile
        fields = '__all__'