requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
astroid==2.7.3
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
channels==2.4.0
channels-redis==2.4.2
charset-normalizer==2.0.5
colorama==0.4.4
constantly==15.1.0
cryptography==35.0.0
daphne==2.5.0
distlib==0.3.3
Django==3.2.7
django-cors-headers==3.8.0
django-storages==1.11.1
djangorestframework==3.12.4
filelock==3.0.12
google==3.0.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==2.0.3
Pillow==8.3.2
platformdirs==2.3.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.10.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
twilio==6.63.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.4.6
wrapt==1.12.1
zipp==3.4.1
zope.interface==5.4.0
